Shellcheck items

+++ scripts/copydocs.sh
@@ -1,8 +1,15 @@
 #!/bin/bash
 
-(cd ../docs; make clean)
-(cd ../docs; make html)
+# Get repo root
+repo_root=$(git rev-parse --show-toplevel 2>/dev/null)
+if [ -z "$repo_root" ]; then
+    echo "Not in a Git repository."
+    exit
+fi
+
+(cd "$repo_root"/docs || exit; make clean)
+(cd "$repo_root"/docs || exit; make html)
 sudo rm -fr /var/www/html/wspr/docs
 sudo mkdir /var/www/html/wspr/docs
-sudo cp -R /home/pi/WsprryPi/docs/_build/html/* /var/www/html/wspr/docs/
+sudo cp -R "$repo_root"/docs/_build/html/* /var/www/html/wspr/docs/
 sudo chown -R www-data:www-data /var/www/html/wspr/docs
